api/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
email-validator
python-multipart
passlib[bcrypt]